pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "synth"
version = "0.1.0"
description = ""
authors = ["Josh Karpel <josh.karpel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.4"
rich = "^13.3.0"
typer = "^0.7.0"
click = "^8.1.3"
pyyaml = "^6.0"

[tool.poetry.scripts]
synth = 'synth.cli:cli'

[tool.black]
line-length = 100
include = "\\.pyi?$"

[tool.isort]
profile = "black"
line_length = 100

[tool.pycln]
all = true
